[IGNORE] Move VariableEditor to plugin-system pkg
