+++ ui/plugin-system/src/components/Variables/VariableEditorForm/VariablePreview.tsx
