+++ ui/plugin-system/src/components/Variables/index.ts
@@ -11,9 +11,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './components';
-export * from './constants';
-export * from './model';
-export * from './runtime';
-export * from './test-utils';
-export * from './utils';
+export * from './variable-model';
+export * from './VariableEditorForm';
