+++ ui/plugin-system/src/components/index.ts
@@ -21,3 +21,4 @@ export * from './PluginKindSelect';
 export * from './PluginRegistry';
 export * from './PluginSpecEditor';
 export * from './TimeSeriesQueryEditor';
+export * from './Variables';
