+++ ui/plugin-system/src/constants/user-interface-text.ts
@@ -11,9 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './components';
-export * from './constants';
-export * from './model';
-export * from './runtime';
-export * from './test-utils';
-export * from './utils';
+export const TOOLTIP_TEXT = {
+  // Variable editor buttons
+  refreshVariableValues: 'Refresh values',
+  copyVariableValues: 'Copy values to clipboard',
+};
